refactor: move channels_v2/connect_channel.py to channels/connect_channel.py
